agents/requirements.txt
google-adk>=1.10.0
google-cloud-logging
google-cloud-storage
python-dotenv
requests
uvicorn
fastapi
nest-asyncio>=1.6.0
httpx
google-genai==1.49.0  # Pinned for File Search API compatibility
